+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Drop default loglevel for initializer & controller down to DEBUG
 * Upgrade controller Docker image to allow for a Docker network per test
 * Run tests in parallel!
+* Provided `--list` flag to the initializer CLI to simply list the tests registered in the suite
 
 # 0.3.1
 * Specify --http-host CLI flag in GetStartCommand to have RPC calls bind to publicIP