+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Run tests in parallel!
 * Removed references to `fiveNode` in testsuite (because these networks are no longer five-node, and the network size is less important than whether it's staking or not)
 * Removed `fiveStakingNodeGetValidatorsTest`, which has been superseded by the RPC workflow test
+* Provided `--list` flag to the initializer CLI to simply list the tests registered in the suite
 
 # 0.3.1
 * Specify --http-host CLI flag in GetStartCommand to have RPC calls bind to publicIP