+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Drop default loglevel for initializer & controller down to DEBUG
 * Upgrade controller Docker image to allow for a Docker network per test
 * Run tests in parallel!
+* Fix bug in RPC workflow test where delegator ID was actually staker ID
 * Removed references to `fiveNode` in testsuite (because these networks are no longer five-node, and the network size is less important than whether it's staking or not)
 * Removed `fiveStakingNodeGetValidatorsTest`, which has been superseded by the RPC workflow test
 * Provided `--list` flag to the initializer CLI to simply list the tests registered in the suite